Fixes #37 - Mark folders as web root folders
New Web Resources > Validation > Web Root Folders property page is
added. It allows zero or more folders from the project to be selected as
web root folders. The settings are stored as persistent property in the
IProject.

The DefaultURIResolver reads the persistent property to find out if any
web root folders are selected for the current project. If yes, then the
existence of files with path containing a leading / are validated
against the web root folders.